Extract chart interface
